move selectbutton out of its original feature since i used it for in this feature
